Drunk update
Originally just showed an html page with the drunk status. Replace with
a better-looking alert box. /drunk/ URL is now just part of the API. A
GET request will return the drunk database as JSON, like the /locations/
URL.

The logic for auto-timeouts for drunk periods thus had to be moved to
the javascript. This is better too, since it's just a call to the API.