Better curves between markers
Instead of using PolyLine objects, which don't make smooth curves, use
an SVG path to draw a curve between each pair of locations. Calculate
the curve using the quadratic Bezier formula with a randomized
curvature. Use the SVG path as the image for a new marker that is
created for each pair.