Improve prevention of overlapping paths
Previously, the curvature of each path was set to a random value. So,
similar paths (e.g. the 2 legs of a round trip) could still randomly
overlap.

Instead, set each path to have a curvature of 0. As new paths are added,
compare the path to previously created paths. Increase the curvature of
the new path for each similar path that is found.

For now, define similar paths as ones with start & end points within 100
miles of each other. This may need to be refined.