Fit bounds within one meridian
By default, google maps will allow LatLngBounds to cross the meridian.
This messes with the SVG paths because if the paths crosses the
meridian, some paths will be drawn on the left and some on the right of
the meridian line.

Avoid this by modifying the LatLngBounds object to be contained within a
single meridian. This effect will still be present as the user pans the
map across the meridian, but this will at least make the default view
look better.