Add plumbing for auto-updating android app
On app resume, request current version for site. Do nothing with the
response for now.

The site will check the version given by app in the POST request. If the
version differs from the current version, the site will return the
current version and a URL to the APK for that version. The app will then
install and run it.