Draw centroid in pixel sizes instead of meters
The GCircle radius size is in meters. Due to map projection, a meter at
the equator is smaller (in pixels) than a meter at a higher latitude.

The effect is that centroids drawn near the equator are smaller than
centroids of the same size drawn elsewhere.

Instead of using GCircle, use the map's projection to convert each
centroid's center coordinates to pixel-based points. Compute circle
paths based on these points, and draw them using GPolygon.